使用canvas压缩图片并保持全图，图片压缩后仍大于1M时会进行提醒。

+++ server/app/service/getUploadAvatarResult.js
@@ -23,7 +23,11 @@ class GetUploadAvatarResultService extends Service {
 			});
 
 			if (selectUserResult.avatar) {
-				fs.unlinkSync(path.join(this.config.baseDir, 'app/public/avatar', selectUserResult.avatar));
+				fs.exists(path.join(this.config.baseDir, 'app/public/avatar', selectUserResult.avatar), flag => {
+					if (flag) {
+						fs.unlinkSync(path.join(this.config.baseDir, 'app/public/avatar', selectUserResult.avatar));
+					}
+				});
 			}
 
 			const updateUserAvatarResult = await this.app.mysql.update('users', {
